% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kfPredict}
\alias{kfPredict}
\title{Kalman Filter Prediction step}
\usage{
kfPredict(x, P, A, Q, B, u)
}
\arguments{
\item{x}{An N x 1 mean state estimate of previous step}

\item{P}{An N x N state covariance of previous step}

\item{A}{(Optional, default idendity) transition matrix of the discrete model}

\item{Q}{(Optional, default zero) process noise of discrete model}

\item{B}{(Optional, default idendity) input effect matrix}

\item{u}{(Optional, default empty) constant input}
}
\value{
A list with two elements
\describe{
  \item{X}{the predicted state mean, and}
  \item{P}{the predicted state covariance.}
}
}
\description{
This function performs the Kalman Filter prediction step
}
\author{
The EKF/UKF Toolbox was written by Simo Särkkä, Jouni Hartikainen,
and Arno Solin.

Dirk Eddelbuettel is porting this package to R and C++, and maintaing it.
}
\seealso{
\link{kfUpdate}, \link{ltiDisc} and 
the documentation for the EKF/UKF toolbox at
\url{http://becs.aalto.fi/en/research/bayes/ekfukf}
}

