% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kfUpdate}
\alias{kfUpdate}
\title{Kalman Filter measurement update step}
\usage{
kfUpdate(x, P, y, H, R)
}
\arguments{
\item{x}{An N x 1 mean state estimate after prediction step}

\item{P}{An N x N state covariance after prediction step}

\item{y}{A D x 1 measurement vector.}

\item{H}{Measurement matrix.}

\item{R}{Measurement noise covariance.}
}
\value{
A list with elements
\describe{
  \item{X}{the update state mean,}
  \item{P}{the update state covariance,}
  \item{K}{the computed Kalman gain,}
  \item{IM}{the mean of the predictive distribution of Y, and}
  \item{IS}{the covariance of the predictive distribution of Y}
}
}
\description{
This function performs the Kalman Filter measurement update step
}
\details{
This functions performs the Kalman Filter measurement update step.
}
\author{
The EKF/UKF Toolbox was written by Simo Särkkä, Jouni Hartikainen,
and Arno Solin.

Dirk Eddelbuettel is porting this package to R and C++, and maintaing it.
}
\seealso{
\link{kfPredict} and 
the documentation for the EKF/UKF toolbox at
\url{http://becs.aalto.fi/en/research/bayes/ekfukf}
}

