% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ltiDisc}
\alias{ltiDisc}
\title{Discretize Linear Time-Invariant ODE}
\usage{
ltiDisc(F, L, Q, dt)
}
\arguments{
\item{F}{An N x N feedback matrix}

\item{L}{(Optional, default idendity) N x L noise effect matrix}

\item{Q}{(Optionalm default zeros) L x L diagonal spectral density}

\item{dt}{(Option, default one) time step}
}
\value{
A list with elements
\describe{
  \item{A}{the transition matrix, and}
  \item{Q}{the discrete process covariance}
}
}
\description{
Discretize Linear Time-Invariant ODE with Gaussian Noise
}
\details{
This function discretizes the linear time-invariant (LTI)
ordinary differential equation (ODE).
}
\author{
The EKF/UKF Toolbox was written by Simo Särkkä, Jouni Hartikainen,
and Arno Solin.

Dirk Eddelbuettel is porting this package to R and C++, and maintaing it.
}
\seealso{
The documentation for the EKF/UKF toolbox at
\url{http://becs.aalto.fi/en/research/bayes/ekfukf}
}

