\name{RcppKalman-package}
\alias{RcppKalman-package}
\alias{RcppKalman}
\docType{package}
\title{
  RcppArmadillo based Kalman filteringKalman Filtering in C++ accessible to R via Rcpp
}
\description{
  This package provides R with a C++ port of some functions from the
  EKF/UKF Toolbox for Matlab by Simo S채rkk채, Jouni Hartikainen, and Arno Solin. 
}
\details{
  TBD
}
\author{
  The EKF/UKF Toolbox for Matlab was written by Simo S채rkk채, Jouni
  Hartikainen, and Arno Solin. Dirk Eddelbuettel wrote this package, and
  maintains it. 
}
\references{
  See \url{http://becs.aalto.fi/en/research/bayes/ekfukf/} for the
  EKF/UKF Toolbox.
}
\keyword{package}
