% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtsSmoother}
\alias{rtsSmoother}
\title{Rauch-Tung-Striebel smoother}
\usage{
rtsSmoother(M, P, A, Q)
}
\arguments{
\item{M}{An N x K matrix of K mean estimates from the Kalman Filter}

\item{P}{An N x N x K cube length K with N x N state covariances matrices 
from the Kalman Filter}

\item{A}{An N x N state transition matrix (or in the more general case a
list of K such matrices; not yet implemented)}

\item{Q}{An N x N noise covariance matrix  (or in the more general case a
list of K such matrices; not yet implemented)}
}
\value{
A list with three elements
\describe{
  \item{SM}{the smoothed mean sequence,}
  \item{SP}{the smooted state covariance sequence,and }
  \item{D}{the smoothed gain sequence.}
}
}
\description{
The function computes the Rauch-Tung-Striebel smoother.
}
\details{
This function implements the Rauch-Tung-Striebel smoother algorithm which
calculate a "smoothed" sequence from the given Kalman filter output sequence
by conditioning all steps to all measurements.
}
\author{
The EKF/UKF Toolbox was written by Simo Särkkä, Jouni Hartikainen,
and Arno Solin.

Dirk Eddelbuettel is porting this package to R and C++, and maintaing it.
}
\seealso{
The documentation for the EKF/UKF toolbox at
\url{http://becs.aalto.fi/en/research/bayes/ekfukf}
}

