% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tfSmoother}
\alias{tfSmoother}
\title{Two-filter Smoother}
\usage{
tfSmoother(M, P, Y, A, Q, H, R, useinf)
}
\arguments{
\item{M}{An N x K matrix of K mean estimates from Kalman filter}

\item{P}{An N x N x K matrix of K state covariances from Kalman Filter}

\item{Y}{A D x K matrix of K measurement sequences}

\item{A}{A N x N state transition matrix.}

\item{Q}{A N x N process noise covariance matrix.}

\item{H}{A D x N measurement matrix.}

\item{R}{A D x D measurement noise covariance.}

\item{useinf}{An optional boolean variable indicating if information
filter should be used (with default \code{true}).}
}
\value{
A list with two elements
\describe{
  \item{M}{the smoothed state mean sequence, and}
  \item{P}{the smoothes state covariance sequence.}
}
}
\description{
This function computes the \sQuote{Two filter-based} Smoother
}
\details{
This function implements the two filter linear smoother which calculates
a \dQuote{smoothed} sequence from the given Kalman filter output sequence
by conditioning all steps to all measurements.
}
\author{
The EKF/UKF Toolbox was written by Simo Särkkä, Jouni Hartikainen,
and Arno Solin.

Dirk Eddelbuettel is porting this package to R and C++, and maintaing it.
}
\seealso{
\link{kfPredict}, \link{kfUpdate}, and 
the documentation for the EKF/UKF toolbox at
\url{http://becs.aalto.fi/en/research/bayes/ekfukf}
}

